+++ NEWS.md
@@ -1,6 +1,7 @@
 # glebrt (development version)
 
 - Refactor `breaks_limits()`
+- Ensure `ggplot2::geom_text()` and `ggplot2::geom_label()` use font set by `set_ggplot_theme()`
 
 # glebrt 0.4.0